WIP: Cleanup semispace
